requirements.txt
flask
rembg[cpu]
onnxruntime
pillow
gunicorn